# librefit-api/openapi.yaml
...
openapi: 3.0.3
info:
  title: librefit-service API
  version: 1.0.0-SNAPSHOT
servers:
- url: http://localhost:8080
  description: Auto generated value
- url: http://0.0.0.0:8080
  description: Auto generated value
paths:
  /api/goals/create:
    post:
      tags:
      - Goals Resource
      operationId: createGoal
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Goal'
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Goal'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
        "403":
          description: Not Allowed
      security:
      - SecurityScheme: []
  /api/goals/delete/{date}/{sequence}:
    delete:
      tags:
      - Goals Resource
      operationId: deleteGoal
      parameters:
      - name: date
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/LocalDate'
      - name: sequence
        in: path
        required: true
        schema:
          format: int64
          type: integer
      responses:
        "200":
          description: OK
          content:
            application/json: {}
        "404":
          description: Not Found
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
        "403":
          description: Not Allowed
      security:
      - SecurityScheme: []
  /api/goals/last:
    get:
      tags:
      - Goals Resource
      operationId: findLastGoal
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Goal'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
        "403":
          description: Not Allowed
      security:
      - SecurityScheme: []
  /api/goals/read/{date}/{sequence}:
    get:
      tags:
      - Goals Resource
      operationId: readGoal
      parameters:
      - name: date
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/LocalDate'
      - name: sequence
        in: path
        required: true
        schema:
          format: int64
          type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Goal'
        "404":
          description: Not Found
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
        "403":
          description: Not Allowed
      security:
      - SecurityScheme: []
  /api/goals/update:
    put:
      tags:
      - Goals Resource
      operationId: updateGoal
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Goal'
      responses:
        "200":
          description: OK
          content:
            application/json: {}
        "404":
          description: Not Found
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
        "403":
          description: Not Allowed
      security:
      - SecurityScheme: []
...
components:
  schemas:
...
    ErrorDescription:
      required:
      - field
      - message
      type: object
      properties:
        field:
          type: string
        message:
          type: string
    ErrorResponse:
      required:
      - errors
      type: object
      properties:
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDescription'
...
    Goal:
      required:
      - added
      - initialWeight
      - targetWeight
      - startDate
      - endDate
      type: object
      properties:
        userId:
          type: string
          allOf:
          - $ref: '#/components/schemas/UUID'
          nullable: true
        added:
          $ref: '#/components/schemas/LocalDate'
        sequence:
          format: int64
          type: integer
          nullable: true
        initialWeight:
          format: float
          minimum: 0
          type: number
        targetWeight:
          format: float
          minimum: 0
          type: number
        startDate:
          $ref: '#/components/schemas/LocalDate'
        endDate:
          $ref: '#/components/schemas/LocalDate'
        targetCalories:
          format: float
          minimum: 0
          type: number
          nullable: true
        maximumCalories:
          format: float
          minimum: 0
          type: number
          nullable: true
        updated:
          type: string
          allOf:
          - $ref: '#/components/schemas/LocalDateTime'
          nullable: true
...
    LocalDate:
      format: date
      type: string
      example: 2022-03-10
    LocalDateTime:
      format: date-time
      type: string
      example: 2022-03-10T12:15:50
...
    UUID:
      format: uuid
      pattern: "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
      type: string
...
  securitySchemes:
    SecurityScheme:
      type: http
      description: Authentication
      scheme: bearer
      bearerFormat: JWT